# packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniDesk Unified API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/automa/workflows/sync/push": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "POST /api/automa/workflows/sync/push\nReceives workflows from the Extension and upserts them into the database + files",
        "operationId": "push_workflows",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushWorkflowsPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workflows synced successfully"
          }
        }
      }
    },
...
    "/api/automa/workflows/{id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "operationId": "get_workflow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workflow ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get workflow by ID"
          },
          "404": {
            "description": "Workflow not found"
          }
        }
      },
      "put": {
        "tags": [
          "workflows"
        ],
        "operationId": "update_workflow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workflow ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workflow updated"
          },
          "404": {
            "description": "Workflow not found"
          }
        }
      },
      "delete": {
        "tags": [
          "workflows"
        ],
        "operationId": "delete_workflow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workflow ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspacePath",
            "in": "query",
            "description": "Workspace path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workflow deleted"
          },
          "404": {
            "description": "Workflow not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PushWorkflowsPayload": {
        "type": "object",
        "description": "Request payload for pushing workflows from the Extension",
        "required": [
          "workflows"
        ],
        "properties": {
          "workflows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workflow"
            }
          }
        }
      },
...
      "Workflow": {
        "type": "object",
        "description": "Workflow definition — mirrors the Automa Extension's workflow JSON structure.",
        "required": [
          "id",
          "name",
          "drawflow",
          "settings"
        ],
        "properties": {
          "connected_table": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "data_columns": {
            "type": [
              "string",
              "null"
            ]
          },
          "delete_source": {
            "type": [
              "string",
              "null"
            ]
          },
          "deleted_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "drawflow": {
            "type": "string"
          },
          "folder_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "global_data": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "is_disabled": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "settings": {
            "type": "string"
          },
          "source": {
            "type": [
              "string",
              "null"
            ]
          },
          "table_data": {
            "type": [
              "string",
              "null"
            ]
          },
          "trigger": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "WorkflowPayload": {
        "type": "object",
        "required": [
          "id",
          "name",
          "drawflow",
          "settings"
        ],
        "properties": {
          "connected_table": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "data_columns": {
            "type": [
              "object",
              "null"
            ]
          },
          "delete_source": {
            "type": [
              "string",
              "null"
            ]
          },
          "deleted_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "drawflow": {
            "type": "object"
          },
          "folder_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "global_data": {
            "type": [
              "object",
              "null"
            ]
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "is_disabled": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "settings": {
            "type": "object"
          },
          "source": {
            "type": [
              "string",
              "null"
            ]
          },
          "table_data": {
            "type": [
              "object",
              "null"
            ]
          },
          "trigger": {
            "type": [
              "object",
              "null"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}